+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 2.0.0
 ------------------
-- Multiline content is supported be beginning lines with `|`.
+- Multiline content is supported by beginning lines with `|`.
 - Static attributes are now denoted with a leading single quote, such as `'msg Hello!` rather than a double quote.
 - Self-closing elements now have proper trailing newline when `keepWhitespace` is enabled.
 - Unanchored attributes (attributes whose indentation don't match a preceding tag) are now properly discarded.